+++ src/main/java/yin/AddEventCommand.java
@@ -1,3 +1,5 @@
+package yin;
+
 import java.time.LocalDateTime;
 
 public class AddEventCommand extends Command {
