+++ src/main/java/yin/Event.java
@@ -1,3 +1,5 @@
+package yin;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
