+++ src/main/java/yin/ExitCommand.java
@@ -1,3 +1,5 @@
+package yin;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
