+++ src/main/java/yin/Storage.java
@@ -1,3 +1,5 @@
+package yin;
+
 import java.io.IOException;
 import java.nio.file.*;
 import java.time.LocalDateTime;
