+++ src/main/java/yin/UnmarkCommand.java
@@ -1,17 +1,19 @@
-public class DeleteCommand extends Command{
+package yin;
+
+public class UnmarkCommand extends  Command {
     private final int index0;
 
-    public DeleteCommand(int index0) {
+    public UnmarkCommand(int index0) {
         this.index0 = index0;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws YinException {
         if (index0 < 0 || index0 >= tasks.size()) {
-            throw new YinException("Invalid index to delete");
+            throw new YinException("Invalid index to unmark");
         }
-        Task removed = tasks.delete(index0);
-        ui.showRemoved(removed, tasks.size());
+        Task task = tasks.unmark(index0);
+        ui.showUnmarked(task);
         storage.save(tasks.asList());
     }
 }
