+++ src/main/java/yin/YinException.java
@@ -1,3 +1,5 @@
+package yin;
+
 public class YinException extends Exception {
     public YinException(String message) {
         super(message);
